fix(apache): Change redirect status code to default of 302
Change of the status code of the RedirectRule from 308 (Permanent
Redirect) to 302 (Found). The status code 308 for a redirect means
that this page is permanently in a new location. This is stored by
the browser and applied directly when visiting again. If a user wants
to change our configuration and, according to our documentation,
disables the Node-RED Proxy, the browser has still saved the
redirection. Although the user has followed all steps, the browser
continues to redirect the user to the "new location". This is
prevented by the status 302, which is the default value for
RewirteRule Fag R.

Link: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/308
Link: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/302
Signed-off-by: Sven Sager <s.sager@kunbus.com>