feat(deb): Declare conflict with other nodered packages
The nodered package from the Raspberry Pi repositories also comes with a
nodered.server unit file. The package must not be installed on the
system at the same time as the revpi-nodered package. The conflict
states that the same files exist in the packages and that remains the
case.

See: https://www.debian.org/doc/debian-policy/ch-relationships.html#conflicting-binary-packages-conflicts
Signed-off-by: Sven Sager <s.sager@kunbus.com>