feat(deb): Replace dependency webstatus with revpi-cert-wizard
In the Bookworm release, the webstatus package will no longer exist, as
it will be replaced by cockpit-revpi. For this reason, it is removed
here as dependency. Since the Apache configuration uses the SSL
certificates from the revpi-cert-wizard, it is defined here as a
dependency. This was previously installed by webstatus.

Signed-off-by: Sven Sager <s.sager@kunbus.com>