fix(apache): Use root of apache default log dir
By writing the logs to the subfolder revpi in the APACHE_LOG_DIR, it
must be created by webstatus and provided with the appropriate
permissions. In addition, a separate Logrotate configuration must be
created. By removing the revpi-webstatus package, these log data must
also be written in the APACHE_LOG_DIR, where the apache2 logrotate
configuration is responsible for the cleanup.

Signed-off-by: Sven Sager <s.sager@kunbus.com>